tests/wpt/web-platform-tests/css/css-pseudo/parsing/target-text.html

<!doctype html>
<meta charset="utf-8">
<title>CSS Pseudo-Elements Test: ::target-text parsing</title>
<link rel="help" href="https://drafts.csswg.org/css-pseudo/#selectordef-target-text">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
  test_valid_selector("::target-text");
  test_valid_selector(".a::target-text");
  test_valid_selector("div ::target-text");
  test_valid_selector("::part(my-part)::target-text");

  test_invalid_selector("::before::target-text");
  test_invalid_selector("::target-text.a");
  test_invalid_selector("::target-text div");
  test_invalid_selector("::target-text::after");
  test_invalid_selector("::target-text:hover");
  test_invalid_selector(":not(::target-text)");
</script>
